fix(backend): use `i64` for `Contact` `INTEGER`s
seems like SQLite doesn't 'really' support anything else, and i don't wanna waste time finding out how to convert it all pretty :p

+++ backend/src/main.rs
@@ -35,8 +35,8 @@ struct Org {
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 struct Contact {
-    pub contact_id: u32,
-    pub org_id: u32,
+    pub contact_id: i64,
+    pub org_id: i64,
     pub first_name: Option<String>,
     pub last_name: Option<String>,
     pub email: Option<String>,
